the trivial part of the actual implementation (#301)
