+++ git-revision/tests/describe/mod.rs
@@ -1,5 +1,6 @@
 use git_object::bstr::ByteSlice;
 use git_revision::describe;
+use git_testtools::hex_to_id;
 use std::borrow::Cow;
 
 mod outcome_format {
@@ -39,7 +40,7 @@ mod outcome_format {
 
 #[test]
 #[ignore]
-fn it_uses_the_ref_the_described_commit_is_on() {
+fn typical_usecases() {
     let dir = git_testtools::scripted_fixture_repo_read_only("make_repo_with_branches.sh").unwrap();
     let repo = git_repository::open(dir).unwrap();
 
@@ -51,17 +52,41 @@ fn it_uses_the_ref_the_described_commit_is_on() {
         hex_len,
         &vec![(commit.id, name.clone())].into_iter().collect(),
     )
-    .unwrap()
-    .expect("find exact match");
+    .unwrap();
+
     assert_eq!(
         res,
-        describe::Outcome {
+        Some(describe::Outcome {
             name,
             id: commit.id,
             depth: 0,
             hex_len,
             long: false,
             dirty_suffix: None
-        }
+        }),
+        "this is an exact match"
+    );
+
+    let name = Cow::Borrowed(b"at-c5".as_bstr());
+    let res = git_revision::describe(
+        &commit.id,
+        hex_len,
+        &vec![(hex_to_id("efd9a841189668f1bab5b8ebade9cd0a1b139a37"), name.clone())]
+            .into_iter()
+            .collect(),
+    )
+    .unwrap();
+
+    assert_eq!(
+        res,
+        Some(describe::Outcome {
+            name,
+            id: commit.id,
+            depth: 3,
+            hex_len,
+            long: false,
+            dirty_suffix: None
+        }),
+        "a match to a tag 1 commit away with 2 commits on the other side of the merge/head"
     );
 }
